Move highlight logic into onBlockLoadedContext
